suppression legacy Module, explosé en plusieurs petits bouts: ModuleInput, Module, ModuleAggregate.
