+++ hesperides-spring/domain/src/main/java/org/hesperides/domain/modules/Techno.java
@@ -1,0 +1,6 @@
+package org.hesperides.domain.modules;
+
+import lombok.Value;
+
+@Value public class Techno {
+}
