+++ hesperides-spring/domain/src/main/java/org/hesperides/domain/modules/commands/CreateModuleCommand.java
@@ -2,6 +2,7 @@ package org.hesperides.domain.modules.commands;
 
 import lombok.Value;
 import org.axonframework.commandhandling.TargetAggregateIdentifier;
+import org.hesperides.domain.modules.Module;
 
 @Value
 public class CreateModuleCommand {
