+++ hesperides-spring/domain/src/main/java/org/hesperides/domain/modules/commands/ModuleAggregate.java
@@ -1,17 +1,13 @@
 package org.hesperides.domain.modules.commands;
 
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import lombok.Value;
 import lombok.extern.slf4j.Slf4j;
 import org.axonframework.commandhandling.CommandHandler;
 import org.axonframework.commandhandling.model.AggregateIdentifier;
 import org.axonframework.eventhandling.EventHandler;
 import org.axonframework.spring.stereotype.Aggregate;
-import org.hesperides.domain.modules.ModuleType;
+import org.hesperides.domain.modules.Module;
 import org.hesperides.domain.modules.events.ModuleCopiedEvent;
 import org.hesperides.domain.modules.events.ModuleCreatedEvent;
-
-import java.net.URI;
 
 import static org.axonframework.commandhandling.model.AggregateLifecycle.apply;
 
@@ -20,32 +16,20 @@ import static org.axonframework.commandhandling.model.AggregateLifecycle.apply;
  */
 @Slf4j
 @Aggregate
-public class Module {
-
-    @Value
-    public static class Key {
-        String name;
-        String version;
-        ModuleType versionType;
-
-        @JsonIgnore
-        public URI getURI() {
-            return URI.create("/rest/modules/" + name + "/" + version + "/" + versionType.name().toLowerCase());
-        }
-    }
+public class ModuleAggregate {
 
     @AggregateIdentifier
-    Key key;
+    Module.Key key;
 
-    public Module() {}
+    public ModuleAggregate() {}
 
     @CommandHandler
-    public Module(CreateModuleCommand command) {
+    public ModuleAggregate(CreateModuleCommand command) {
         apply(new ModuleCreatedEvent(command.getModuleKey()));
     }
 
     @CommandHandler
-    public Module(CopyModuleCommand command) {
+    public ModuleAggregate(CopyModuleCommand command) {
         apply(new ModuleCopiedEvent(command.getModuleKey(), command.getSourceModuleKey()));
     }
