+++ hesperides-spring/domain/src/main/java/org/hesperides/domain/modules/queries/ModuleByIdQuery.java
@@ -1,7 +1,7 @@
 package org.hesperides.domain.modules.queries;
 
 import lombok.Value;
-import org.hesperides.domain.modules.commands.Module;
+import org.hesperides.domain.modules.Module;
 
 /**
  * recherche un module par sa clé.
