+++ hesperides-spring/domain/src/test/java/org/hesperides/domain/modules/ModuleAggregateTest.java
@@ -4,19 +4,19 @@ import org.axonframework.test.aggregate.AggregateTestFixture;
 import org.axonframework.test.aggregate.FixtureConfiguration;
 import org.hesperides.domain.modules.commands.CopyModuleCommand;
 import org.hesperides.domain.modules.commands.CreateModuleCommand;
-import org.hesperides.domain.modules.commands.Module;
+import org.hesperides.domain.modules.commands.ModuleAggregate;
 import org.hesperides.domain.modules.events.ModuleCopiedEvent;
 import org.hesperides.domain.modules.events.ModuleCreatedEvent;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-class ModuleTest {
+class ModuleAggregateTest {
 
-    private FixtureConfiguration<Module> fixture;
+    private FixtureConfiguration<ModuleAggregate> fixture;
 
     @BeforeEach
     void setUp() throws Exception {
-        fixture = new AggregateTestFixture<>(Module.class);
+        fixture = new AggregateTestFixture<>(ModuleAggregate.class);
     }
 
     @Test
@@ -42,7 +42,7 @@ class ModuleTest {
 //
 //    @Test
 //    void given_an_existing_module_release_command_should_change_version() {
-//        Module.Key id = new Module.Key("module_test","123");
+//        ModuleAggregate.Key id = new ModuleAggregate.Key("module_test","123");
 //
 //        fixture.given(new ModuleCreatedEvent(id))
 //                .when(new ReleaseModuleCommand(id,"123"))
