+++ hesperides-spring/infrastructure/src/main/java/org/hesperides/infrastructure/modules/local/LocalModuleRepository.java
@@ -4,10 +4,9 @@ import com.google.common.collect.ImmutableList;
 import com.google.common.collect.Maps;
 import lombok.extern.slf4j.Slf4j;
 import org.axonframework.eventsourcing.EventSourcingHandler;
-import org.axonframework.messaging.MetaData;
 import org.axonframework.queryhandling.QueryHandler;
+import org.hesperides.domain.modules.Module;
 import org.hesperides.domain.modules.ModuleType;
-import org.hesperides.domain.modules.commands.Module;
 import org.hesperides.domain.modules.events.ModuleCreatedEvent;
 import org.hesperides.domain.modules.queries.ModuleByIdQuery;
 import org.hesperides.domain.modules.queries.ModuleView;
