--- hesperides-spring/presentation/src/main/java/org/hesperides/presentation/legacydtos/HesperidesException.java
@@ -1,31 +1,0 @@
-/*
- *
- *  * This file is part of the Hesperides distribution.
- *  * (https://github.com/voyages-sncf-technologies/hesperides)
- *  * Copyright (c) 2016 VSCT.
- *  *
- *  * Hesperides is free software: you can redistribute it and/or modify
- *  * it under the terms of the GNU General Public License as
- *  * published by the Free Software Foundation, version 3.
- *  *
- *  * Hesperides is distributed in the hope that it will be useful, but
- *  * WITHOUT ANY WARRANTY; without even the implied warranty of
- *  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
- *  * General Public License for more details.
- *  *
- *  * You should have received a copy of the GNU General Public License
- *  * along with this program. If not, see <http://www.gnu.org/licenses/>.
- *
- *
- */
-
-package org.hesperides.presentation.legacydtos;
-
-/**
- * Created by william_montaz on 29/04/2015.
- */
-public class HesperidesException extends RuntimeException {
-    public HesperidesException(String message) {
-        super(message);
-    }
-}
