--- hesperides-spring/presentation/src/main/java/org/hesperides/presentation/legacydtos/package-info.java
@@ -1,5 +1,0 @@
-/**
- * ce package regroupe les classes issues du legacy qu'on utilise en input de l'api REST
- * on pourra les bouger plus tard.
- */
-package org.hesperides.presentation.legacydtos;
